.circleci: Replace more `docker cp` -> volume mounts
`docker cp` is not the recommended way of moving artifacts in and out of
containers and we should prefer to use volume mounting wherever possible
so that the ci machines do not have to do a tar and copy.

Includes a TODO to cleanup the gradle builds that is a bit too
complicated for this commit.

Signed-off-by: Eli Uriegas <eliuriegas@fb.com>