.github: Split flaky rules / merge rules reads
This diff removes the `flaky_rules_location_url` from the
merge_rules.yml due to issues with internal `github-export-checks`
attempting to read this value and failing. See
https://fburl.com/sandcastle/7e3ks8jo as an example of a failed
`github-export-checks` with this configuration value.

merge_rules.yml needs to have a consistent format in order for internal
`github-export-checks` to work properly since the code that runs that
check is always behind what is on Github master making races like this
possible.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>